refactor(api): normalize blank default kill chain to null and move SecurityPlatformSimpleOutput to api package (#6958)
