+++ openaev-model/src/main/java/io/openaev/database/model/Scenario.java
@@ -167,12 +167,20 @@ public class Scenario extends ModelBehaviour implements GrantableBase, TenantBas
 
   /**
    * Kill chain (by name, e.g. "mitre-attack") displayed first in the overview's kill chain results.
-   * Null means automatic (ATT&CK first). A user's own selection, remembered in local storage, still
-   * overrides this default.
+   * Null means automatic (ATT&CK first); blank input (the UI's "Automatic" option) is normalized to
+   * null on write. A user's own selection, remembered in local storage, still overrides this
+   * default.
    */
+  @Setter(NONE)
   @Column(name = "scenario_default_kill_chain")
   @JsonProperty("scenario_default_kill_chain")
   private String defaultKillChain;
+
+  public void setDefaultKillChain(String defaultKillChain) {
+    // The UI sends "" for "Automatic": normalize so null is the only automatic marker in DB.
+    this.defaultKillChain =
+        (defaultKillChain == null || defaultKillChain.isBlank()) ? null : defaultKillChain;
+  }
 
   /**
    * Whether the expectation-drift warning was dismissed for this scenario (the drifted expectations
