+++ openaev-model/src/main/java/io/openaev/database/raw/RawSimulationIndexing.java
@@ -67,7 +67,8 @@ public interface RawSimulationIndexing extends RawTenant {
   /**
    * Returns the kill chain displayed first in the overview kill chain results.
    *
-   * @return the default kill chain name (e.g., "mitre-attack"), or null for automatic
+   * @return the default kill chain name (e.g., "mitre-attack"), or null for automatic (blank input
+   *     is normalized to null on write, so null is the only automatic marker)
    */
   String getExercise_default_kill_chain();
